h3: add fin flag to HTTP/3 Header Events
Previously, when `quiche::h3::Connection::poll()`
returned a `Headers` event, there was no way to tell
if it simultaneously closed the indicated stream.
Instead, `poll()` would need to be called again
until the `Finished` event occurred.

With this change a boolean `fin` flag is returned
with the Event, which allows for some more
optimised work flows.

This is changes the Rust API, the Headers event
is changes from a `Vec<Header>` to a struct.

This changes the FFI, the `quiche_h3_event_for_each_header`
method takes an additional parameter.